+++ crates/cli/src/commands/outdated.rs
@@ -1,7 +1,8 @@
 use crate::error::ProtoCliError;
+use crate::helpers::ProtoResource;
 use clap::Args;
 use miette::IntoDiagnostic;
-use proto_core::{load_tool, ToolsConfig, UnresolvedVersionSpec, VersionSpec};
+use proto_core::{ProtoConfig, UnresolvedVersionSpec, VersionSpec};
 use serde::Serialize;
 use starbase::system;
 use starbase_styles::color::{self, OwoStyle};
@@ -11,6 +12,9 @@ use tracing::{debug, info};
 
 #[derive(Args, Clone, Debug)]
 pub struct OutdatedArgs {
+    #[arg(long, help = "Include versions in global .prototools")]
+    include_global: bool,
+
     #[arg(long, help = "Print the list in JSON format")]
     json: bool,
 
@@ -19,6 +23,9 @@ pub struct OutdatedArgs {
         help = "Check for latest available version ignoring requirements and ranges"
     )]
     latest: bool,
+
+    #[arg(long, help = "Only check versions in local .prototools")]
+    only_local: bool,
 
     #[arg(long, help = "Update and write the versions to the local .prototools")]
     update: bool,
@@ -33,27 +40,31 @@ pub struct OutdatedItem {
 }
 
 #[system]
-pub async fn outdated(args: ArgsRef<OutdatedArgs>) {
-    let mut tools_config = ToolsConfig::load_closest()?;
-    let initial_version = UnresolvedVersionSpec::default(); // latest
+pub async fn outdated(args: ArgsRef<OutdatedArgs>, proto: ResourceRef<ProtoResource>) {
+    let manager = proto.env.load_config_manager()?;
 
-    if tools_config.tools.is_empty() {
-        return Err(ProtoCliError::NoConfiguredTools {
-            path: tools_config.path,
-        }
-        .into());
+    let config = if args.only_local {
+        manager.get_local_config()?
+    } else if args.include_global {
+        manager.get_merged_config()?
+    } else {
+        manager.get_merged_config_without_global()?
+    };
+
+    if config.versions.is_empty() {
+        return Err(ProtoCliError::NoConfiguredTools.into());
     }
 
     if !args.json {
         info!("Checking for newer versions...");
-        info!("Loading {}", color::path(&tools_config.path));
     }
 
     let mut items = HashMap::new();
     let mut tool_versions = HashMap::new();
+    let initial_version = UnresolvedVersionSpec::default(); // latest
 
-    for (tool_id, config_version) in &tools_config.tools {
-        let mut tool = load_tool(tool_id).await?;
+    for (tool_id, config_version) in &config.versions {
+        let mut tool = proto.load_tool(tool_id).await?;
         tool.disable_caching();
 
         debug!("Checking {}", tool.get_name());
@@ -61,7 +72,8 @@ pub async fn outdated(args: ArgsRef<OutdatedArgs>) {
         let mut comments = vec![];
         let versions = tool.load_version_resolver(&initial_version).await?;
         let current_version = versions.resolve(config_version)?;
-        let is_latest = args.latest || matches!(config_version, UnresolvedVersionSpec::Version(_));
+        let check_latest =
+            args.latest || matches!(config_version, UnresolvedVersionSpec::Version(_));
 
         comments.push(format!(
             "current version {} {}",
@@ -69,29 +81,46 @@ pub async fn outdated(args: ArgsRef<OutdatedArgs>) {
             color::muted_light(format!("(via {})", config_version))
         ));
 
-        let newer_version = versions.resolve_without_manifest(if is_latest {
+        let newer_version = versions.resolve_without_manifest(if check_latest {
             &initial_version // latest alias
         } else {
             config_version // req, range, etc
         })?;
 
-        comments.push(format!(
-            "{} {}",
-            if is_latest {
-                "latest version"
+        let mut is_outdated = false;
+        let mut is_on_latest = false;
+
+        if let (VersionSpec::Version(a), VersionSpec::Version(b)) =
+            (&current_version, &newer_version)
+        {
+            #[allow(clippy::comparison_chain)]
+            if b > a {
+                is_outdated = true;
+            } else if b == a {
+                is_on_latest = true;
+            }
+        }
+
+        if is_on_latest {
+            comments.push(if check_latest {
+                "on the latest version".into()
             } else {
-                "newer version"
-            },
-            color::symbol(newer_version.to_string())
-        ));
+                "on the newest version".into()
+            });
+        } else {
+            comments.push(format!(
+                "{} {}",
+                if check_latest {
+                    "latest version"
+                } else {
+                    "newer version"
+                },
+                color::symbol(newer_version.to_string())
+            ));
 
-        let is_outdated = match (&current_version, &newer_version) {
-            (VersionSpec::Version(a), VersionSpec::Version(b)) => b > a,
-            _ => false,
-        };
-
-        if is_outdated {
-            comments.push(color::success("update available!"));
+            if is_outdated {
+                comments.push(color::success("update available!"));
+            }
         }
 
         if args.update {
@@ -102,7 +131,7 @@ pub async fn outdated(args: ArgsRef<OutdatedArgs>) {
             items.insert(
                 tool.id,
                 OutdatedItem {
-                    is_latest,
+                    is_latest: check_latest,
                     version_config: config_version.to_owned(),
                     current_version,
                     newer_version,
@@ -119,8 +148,12 @@ pub async fn outdated(args: ArgsRef<OutdatedArgs>) {
     }
 
     if args.update {
-        tools_config.tools.extend(tool_versions);
-        tools_config.save()?;
+        ProtoConfig::update(&proto.env.cwd, |config| {
+            config
+                .versions
+                .get_or_insert(Default::default())
+                .extend(tool_versions);
+        })?;
     }
 
     if args.json {