+++ crates/core/src/proto.rs
@@ -1,11 +1,13 @@
 use crate::helpers::{get_home_dir, get_proto_home, is_offline};
-use crate::user_config::UserConfig;
+use crate::proto_config::{ProtoConfig, ProtoConfigManager};
+use crate::{ProtoConfigFile, PROTO_CONFIG_NAME};
 use once_cell::sync::OnceCell;
 use std::collections::BTreeMap;
 use std::env;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use warpgate::{create_http_client_with_options, PluginLoader};
+use tracing::debug;
+use warpgate::PluginLoader;
 
 #[derive(Clone)]
 pub struct ProtoEnvironment {
@@ -18,8 +20,9 @@ pub struct ProtoEnvironment {
     pub home: PathBuf, // ~
     pub root: PathBuf, // ~/.proto
 
-    client: Arc<OnceCell<reqwest::Client>>,
-    loader: Arc<OnceCell<PluginLoader>>,
+    config_manager: Arc<OnceCell<ProtoConfigManager>>,
+    plugin_loader: Arc<OnceCell<PluginLoader>>,
+    test_mode: bool,
 }
 
 impl ProtoEnvironment {
@@ -31,11 +34,14 @@ impl ProtoEnvironment {
         let mut env = Self::from(sandbox.join(".proto")).unwrap();
         env.cwd = sandbox.to_path_buf();
         env.home = sandbox.join(".home");
+        env.test_mode = true;
         env
     }
 
     pub fn from<P: AsRef<Path>>(root: P) -> miette::Result<Self> {
         let root = root.as_ref();
+
+        debug!(store = ?root, "Creating proto environment, detecting store");
 
         Ok(ProtoEnvironment {
             bin_dir: root.join("bin"),
@@ -46,26 +52,30 @@ impl ProtoEnvironment {
             tools_dir: root.join("tools"),
             home: get_home_dir()?,
             root: root.to_owned(),
-            client: Arc::new(OnceCell::new()),
-            loader: Arc::new(OnceCell::new()),
+            config_manager: Arc::new(OnceCell::new()),
+            plugin_loader: Arc::new(OnceCell::new()),
+            test_mode: false,
         })
     }
 
-    pub fn get_http_client(&self) -> miette::Result<&reqwest::Client> {
-        self.client.get_or_try_init(|| {
-            let user_config = UserConfig::load()?;
-            let client = create_http_client_with_options(user_config.http)?;
-
-            Ok(client)
-        })
+    pub fn get_config_dir(&self, global: bool) -> &Path {
+        if global {
+            &self.root
+        } else {
+            &self.cwd
+        }
     }
 
-    pub fn get_plugin_loader(&self) -> &PluginLoader {
-        self.loader.get_or_init(|| {
+    pub fn get_plugin_loader(&self) -> miette::Result<&PluginLoader> {
+        let config = self.load_config()?;
+
+        self.plugin_loader.get_or_try_init(|| {
             let mut loader = PluginLoader::new(&self.plugins_dir, &self.temp_dir);
+            loader.set_client_options(&config.settings.http);
             loader.set_offline_checker(is_offline);
             loader.set_seed(env!("CARGO_PKG_VERSION"));
-            loader
+
+            Ok(loader)
         })
     }
 
@@ -77,8 +87,34 @@ impl ProtoEnvironment {
         ])
     }
 
-    pub fn load_user_config(&self) -> miette::Result<UserConfig> {
-        UserConfig::load_from(&self.root)
+    pub fn load_config(&self) -> miette::Result<&ProtoConfig> {
+        self.load_config_manager()?.get_merged_config()
+    }
+
+    pub fn load_config_manager(&self) -> miette::Result<&ProtoConfigManager> {
+        self.config_manager.get_or_try_init(|| {
+            // Don't traverse passed the home directory,
+            // but only if working directory is within it!
+            let end_dir = if self.cwd.starts_with(&self.home) {
+                Some(self.home.as_path())
+            } else {
+                None
+            };
+
+            let mut manager = ProtoConfigManager::load(&self.cwd, end_dir)?;
+
+            // Always load the proto home/root config last
+            let path = self.root.join(PROTO_CONFIG_NAME);
+
+            manager.files.push(ProtoConfigFile {
+                exists: path.exists(),
+                global: true,
+                path,
+                config: ProtoConfig::load_from(&self.root, true)?,
+            });
+
+            Ok(manager)
+        })
     }
 }